Hamit/requirements.txt
openai
youtube_transcript_api
python-dotenv
gradio
